Moved errors to error module & added RpcError into client.rs
