+++ workspaces/src/rpc/client.rs
@@ -22,8 +22,8 @@ use near_primitives::views::{
     QueryRequest, StatusResponse,
 };
 
-use crate::error::{Error, RpcErrorKind};
-use crate::result::{Result, ViewResultDetails};
+use crate::error::{RpcError, RpcErrorKind};
+use crate::result::ViewResultDetails;
 use crate::rpc::tool;
 use crate::types::{AccountId, InMemorySigner, PublicKey, Signer};
 
@@ -116,7 +116,7 @@ impl Client {
         signer: &InMemorySigner,
         receiver_id: &AccountId,
         action: Action,
-    ) -> Result<FinalExecutionOutcomeView> {
+    ) -> Result<FinalExecutionOutcomeView, RpcError> {
         send_batch_tx_and_retry(self, signer, receiver_id, vec![action]).await
     }
 
@@ -128,7 +128,7 @@ impl Client {
         args: Vec<u8>,
         gas: Gas,
         deposit: Balance,
-    ) -> Result<FinalExecutionOutcomeView> {
+    ) -> Result<FinalExecutionOutcomeView, RpcError> {
         self.send_tx_and_retry(
             signer,
             contract_id,
@@ -148,7 +148,7 @@ impl Client {
         contract_id: AccountId,
         method_name: String,
         args: Vec<u8>,
-    ) -> Result<ViewResultDetails> {
+    ) -> Result<ViewResultDetails, RpcError> {
         let query_resp = self
             .query(&RpcQueryRequest {
                 block_reference: Finality::None.into(), // Optimisitic query
@@ -174,7 +174,7 @@ impl Client {
         contract_id: AccountId,
         prefix: Option<&[u8]>,
         block_id: Option<BlockId>,
-    ) -> Result<HashMap<Vec<u8>, Vec<u8>>> {
+    ) -> crate::result::Result<HashMap<Vec<u8>, Vec<u8>>> {
         let block_reference = block_id
             .map(Into::into)
             .unwrap_or_else(|| Finality::None.into());
@@ -202,7 +202,7 @@ impl Client {
         &self,
         account_id: AccountId,
         block_id: Option<BlockId>,
-    ) -> Result<AccountView> {
+    ) -> Result<AccountView, RpcError> {
         let block_reference = block_id
             .map(Into::into)
             .unwrap_or_else(|| Finality::None.into());
@@ -227,7 +227,7 @@ impl Client {
         &self,
         account_id: AccountId,
         block_id: Option<BlockId>,
-    ) -> Result<ContractCodeView> {
+    ) -> Result<ContractCodeView, RpcError> {
         let block_reference = block_id
             .map(Into::into)
             .unwrap_or_else(|| Finality::None.into());
@@ -248,7 +248,10 @@ impl Client {
         }
     }
 
-    pub(crate) async fn view_block(&self, block_id: Option<BlockId>) -> Result<BlockView> {
+    pub(crate) async fn view_block(
+        &self,
+        block_id: Option<BlockId>,
+    ) -> Result<BlockView, RpcError> {
         let block_reference = block_id
             .map(Into::into)
             .unwrap_or_else(|| Finality::None.into());
@@ -266,7 +269,7 @@ impl Client {
         signer: &InMemorySigner,
         contract_id: &AccountId,
         wasm: Vec<u8>,
-    ) -> Result<FinalExecutionOutcomeView> {
+    ) -> Result<FinalExecutionOutcomeView, RpcError> {
         self.send_tx_and_retry(
             signer,
             contract_id,
@@ -281,7 +284,7 @@ impl Client {
         signer: &InMemorySigner,
         receiver_id: &AccountId,
         amount_yocto: Balance,
-    ) -> Result<FinalExecutionOutcomeView> {
+    ) -> Result<FinalExecutionOutcomeView, RpcError> {
         self.send_tx_and_retry(
             signer,
             receiver_id,
@@ -299,7 +302,7 @@ impl Client {
         new_account_id: &AccountId,
         new_account_pk: PublicKey,
         amount: Balance,
-    ) -> Result<FinalExecutionOutcomeView> {
+    ) -> Result<FinalExecutionOutcomeView, RpcError> {
         send_batch_tx_and_retry(
             self,
             signer,
@@ -327,7 +330,7 @@ impl Client {
         new_account_pk: PublicKey,
         amount: Balance,
         code: Vec<u8>,
-    ) -> Result<FinalExecutionOutcomeView> {
+    ) -> Result<FinalExecutionOutcomeView, RpcError> {
         send_batch_tx_and_retry(
             self,
             signer,
@@ -355,7 +358,7 @@ impl Client {
         signer: &InMemorySigner,
         account_id: &AccountId,
         beneficiary_id: &AccountId,
-    ) -> Result<FinalExecutionOutcomeView> {
+    ) -> Result<FinalExecutionOutcomeView, RpcError> {
         let beneficiary_id = beneficiary_id.to_owned();
         self.send_tx_and_retry(
             signer,
@@ -378,13 +381,14 @@ impl Client {
         result
     }
 
-    pub(crate) async fn wait_for_rpc(&self) -> Result<()> {
+    pub(crate) async fn wait_for_rpc(&self) -> Result<(), RpcError> {
         let timeout_secs = match std::env::var("NEAR_RPC_TIMEOUT_SECS") {
-            Ok(secs) => secs.parse::<usize>().map_err(|e| Error::Other(e.into()))?,
+            // hard fail on not being able to parse the env var, since this isn't something
+            // the user should handle with the library.
+            Ok(secs) => secs.parse::<usize>().unwrap(),
             Err(_) => 10,
         };
 
-        // TODO: proper way to handle JsonRpcError inside workspace error
         let retry_strategy =
             std::iter::repeat_with(|| Duration::from_millis(500)).take(2 * timeout_secs);
         Retry::spawn(retry_strategy, || async { self.status().await })
@@ -395,7 +399,7 @@ impl Client {
                     self.rpc_addr, timeout_secs, e
                 )));
 
-                Into::<Error>::into(err)
+                Into::<RpcError>::into(err)
             })?;
         Ok(())
     }
@@ -405,7 +409,7 @@ pub(crate) async fn access_key(
     client: &Client,
     account_id: near_primitives::account::id::AccountId,
     public_key: near_crypto::PublicKey,
-) -> Result<(AccessKeyView, CryptoHash)> {
+) -> Result<(AccessKeyView, CryptoHash), RpcError> {
     let query_resp = client
         .query(&methods::query::RpcQueryRequest {
             block_reference: Finality::None.into(),
@@ -440,7 +444,7 @@ where
 pub(crate) async fn send_tx(
     client: &Client,
     tx: SignedTransaction,
-) -> Result<FinalExecutionOutcomeView> {
+) -> Result<FinalExecutionOutcomeView, RpcError> {
     client
         .query_broadcast_tx(&methods::broadcast_tx_commit::RpcBroadcastTxCommitRequest {
             signed_transaction: tx,
@@ -453,10 +457,10 @@ pub(crate) async fn send_tx(
 pub(crate) async fn send_tx_and_retry<T, F>(
     client: &Client,
     task: F,
-) -> Result<FinalExecutionOutcomeView>
+) -> Result<FinalExecutionOutcomeView, RpcError>
 where
     F: Fn() -> T,
-    T: core::future::Future<Output = Result<SignedTransaction>>,
+    T: core::future::Future<Output = Result<SignedTransaction, RpcError>>,
 {
     retry(|| async { send_tx(client, task().await?).await }).await
 }
@@ -466,7 +470,7 @@ pub(crate) async fn send_batch_tx_and_retry(
     signer: &InMemorySigner,
     receiver_id: &AccountId,
     actions: Vec<Action>,
-) -> Result<FinalExecutionOutcomeView, Error> {
+) -> Result<FinalExecutionOutcomeView, RpcError> {
     send_tx_and_retry(client, || async {
         let (AccessKeyView { nonce, .. }, block_hash) = access_key(
             client,
